Fix checkGenericClassInNonGenericObjectType PHPStan errors
This commit removes the checkGenericClassInNonGenericObjectType flag
from PHPStan configuration files and fixes all associated errors in our
code base.

[MAILPOET-3236]

+++ lib/Segments/SegmentsRepository.php
@@ -111,8 +111,10 @@ class SegmentsRepository extends Repository {
     // We want to remove redundant filters before update
     while ($segment->getDynamicFilters()->count() > count($filtersData)) {
       $filterEntity = $segment->getDynamicFilters()->last();
-      $segment->getDynamicFilters()->removeElement($filterEntity);
-      $this->entityManager->remove($filterEntity);
+      if ($filterEntity) {
+        $segment->getDynamicFilters()->removeElement($filterEntity);
+        $this->entityManager->remove($filterEntity);
+      }
     }
     foreach ($filtersData as $key => $filterData) {
       if ($filterData instanceof DynamicSegmentFilterData) {
